+++ tests/core/test_utils.py
@@ -30,3 +30,15 @@ def test_dependency_debug():
                                   dependencies=[2])
     assert ref != DependencyDebug(info="info", singleton=True, wired=[1],
                                   dependencies=[20])
+
+
+def test_dependency_hash_eq():
+    d1 = Dependency('same')
+    d2 = Dependency('same')
+    x = Dependency('different')
+
+    assert hash(d1) == hash(d2)
+    assert d1 == d2
+
+    assert hash(d1) != hash(x)
+    assert d1 != x